requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
flask-limiter>=3.5.0
PyJWT>=2.8.0
bcrypt>=4.1.0
scikit-learn>=1.4.0
shap>=0.44.0
joblib>=1.3.0
python-dotenv>=1.0.0
resend>=2.0.0
numpy>=1.26.0
scipy>=1.12.0
# MySQL support (pure Python — no C compiler needed)
PyMySQL>=1.1.0
cryptography>=42.0.0
